refactor replace user contextAPI with user redux reducer
