+++ src/store/user.tsx
@@ -4,6 +4,13 @@ export const USER_ACTION_TYPES = {
   SET_CURRENT_USER: 'SET_CURRENT_USER',
 };
 
+export const setCurrentUser = (user: UserAuth | null) => {
+  return {
+    type: USER_ACTION_TYPES.SET_CURRENT_USER,
+    payload: user,
+  };
+};
+
 const INITIAL_STATE = {
   currentUser: {
     uid: '',
